[#63] Renamed all files to index.js where necessary
