+++ src/components/header/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { connect } from 'react-redux';
 import { withRouter } from 'react-router-dom';
 
-import { getTeams } from 'actions/actions';
+import { getTeams } from 'actions';
 
 import DropdownMenu from './dropdownMenu';
 import UserLogin from './userLogin';
